feat: Update asset copying and animation handling
This commit updates the CMakeLists.txt to copy the assets directory
tothe build directory during the build process. It also refactors
theanimation handling code to use an unordered_multiset instead of a
vector,which fixes a bug where the animations vector was being modified
whilebeing iterated over. The paths to assets in the code have also
beenupdated to reflect the new location of the assets directory.