feat: Add menu time and max score to game
This commit introduces two new features to the game. The first is
theaddition of a menu time, which is subtracted from the total game
timeto give a more accurate representation of the time spent playing
thegame. The second feature is the addition of a max score, which
keepstrack of the highest score achieved in the game. These features
enhancethe user experience by providing more detailed game statistics.