feat(GUI): Refactor Button class and add BigButton class
The Button class has been refactored to be more generic and a
newBigButton class has been added. The BigButton class is used
fordifferent game states. The GUI texture is now loaded once and
passedto the button classes. The game state is also passed to the
buttonclasses to allow them to change the game state on click. The GUI
menustate has been removed as it is no longer needed.